+++ test/dummy/app/models/group.rb
@@ -1,2 +1,3 @@
 class Group < ApplicationRecord
+  has_many :users
 end